llm_chat_interface/requirements.txt
gradio
google-generativeai
requests
pandas
python-dotenv